+++ src/main/java/faang/school/godbless/BJS2_1219/House.java
@@ -5,6 +5,9 @@ import lombok.NoArgsConstructor;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
 
 @Data
 @NoArgsConstructor
@@ -12,17 +15,41 @@ public class House {
     private List<Food> collectedFood = new ArrayList<>();
     private List<Room> rooms;
 
+    public static void main(String[] args) {
+        House house = new House();
+        house.initialize();
+
+        ScheduledExecutorService executor = Executors.newScheduledThreadPool(5);
+
+        for (int i = 0; i < house.rooms.size() / 2; i++) {
+            executor.schedule(house::collectFood, i * 30L, TimeUnit.SECONDS);
+        }
+        executor.shutdown();
+
+        try {
+            executor.awaitTermination(5, TimeUnit.MINUTES);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+
+        System.out.println("Collected food: " + house.getCollectedFood());
+        System.out.println("Food collected!");
+    }
 
     public void collectFood() {
         int count = 0;
         for (Room room : rooms) {
-            if (count >= 2 || room.getFoods().isEmpty()) {
+            if (count >= 2) {
+                break;
+            }
+            if (room.getFoods() == null || room.getFoods().isEmpty()) {
                 continue;
             }
             collectedFood.addAll(room.getFoods());
             room.getFoods().clear();
             count++;
         }
+        System.out.println("In thread " + Thread.currentThread() + " collected food: " + collectedFood);
     }
 
     public void initialize() {
