--- src/main/java/faang/school/godbless/BJS2_1219/Main.java
@@ -1,30 +1,0 @@
-package faang.school.godbless.BJS2_1219;
-
-import java.util.concurrent.Executors;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
-
-public class Main {
-    public static void main(String[] args) {
-
-        House house = new House();
-        house.initialize();
-
-        ScheduledExecutorService executor = Executors.newScheduledThreadPool(5);
-
-        for (int i = 0; i < 5; i++) {
-            executor.schedule(house::collectFood, i * 2, TimeUnit.SECONDS);
-        }
-        executor.shutdown();
-
-        try {
-            executor.awaitTermination(5, TimeUnit.MINUTES);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-
-        System.out.println("Collected food: " + house.getCollectedFood());
-
-        System.out.println("Food collected!");
-    }
-}
