+++ src/main/java/faang/school/godbless/BJS2_1219/Room.java
@@ -2,7 +2,6 @@ package faang.school.godbless.BJS2_1219;
 
 import lombok.Data;
 
-import java.util.ArrayList;
 import java.util.List;
